+++ src/FLEx-ChorusPlugin/Infrastructure/FlexBridgeSychronizerAdjunct.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
@@ -7,8 +8,8 @@ using Chorus.VcsDrivers.Mercurial;
 using Chorus.sync;
 using FLEx_ChorusPlugin.Infrastructure.DomainServices;
 using FLEx_ChorusPlugin.Properties;
-using Microsoft.Win32;
 using Palaso.Progress;
+using TriboroughBridge_ChorusPlugin.Properties;
 
 namespace FLEx_ChorusPlugin.Infrastructure
 {
@@ -17,17 +18,14 @@ namespace FLEx_ChorusPlugin.Infrastructure
 		private readonly string _fwdataPathname;
 		private readonly bool _writeVerbose;
 		private bool _needToNestMainFile = true;
-		private string _fixitPathname;
+		private readonly string _fixitPathname;
 
-		internal FlexBridgeSychronizerAdjunct(string fwdataPathname, string fixitPathname)
-			: this(fwdataPathname, true)
+		internal FlexBridgeSychronizerAdjunct(string fwdataPathname, string fixitPathname, bool writeVerbose)
 		{
-			_fixitPathname = fixitPathname;
-		}
-
-		internal FlexBridgeSychronizerAdjunct(string fwdataPathname, bool writeVerbose)
-		{
+			if (!File.Exists(fixitPathname))
+				throw new InvalidOperationException("The FLEx 'fix it' program was not found.");
 			_fwdataPathname = fwdataPathname;
+			_fixitPathname = fixitPathname;
 			_writeVerbose = writeVerbose;
 		}
 
@@ -84,7 +82,7 @@ namespace FLEx_ChorusPlugin.Infrastructure
 			progress.WriteMessage("Checking project for merge problems");
 			if (RunFixFwData(progress))
 			{
-				progress.WriteWarning("Fixed some merge problems");
+				progress.WriteMessage("Ran fix-up utility after merge.");
 				FLExProjectSplitter.PushHumptyOffTheWall(progress, _writeVerbose, _fwdataPathname);
 			}
 		}
@@ -95,39 +93,10 @@ namespace FLEx_ChorusPlugin.Infrastructure
 		/// <returns>true if problems were fixed</returns>
 		private bool RunFixFwData(IProgress progress)
 		{
-			if (string.IsNullOrEmpty(_fixitPathname))
-			{
-				// This shoudl now only be called by the stand alone option, as the Felx driven side will feed the pathname as a '-f' option.
-				const string fixerName = @"FixFwData.exe";
-				// This should work on user machines.
-				var codeDir = (string)Registry.LocalMachine.OpenSubKey(@"Software\SIL\FieldWorks\7.0\").GetValue(@"RootCodeDir");
-				_fixitPathname = Path.Combine(codeDir, fixerName);
-				if (!File.Exists(_fixitPathname))
-				{
-					// This should work on developer machines.
-					codeDir = (string)Registry.LocalMachine.OpenSubKey(@"Software\SIL\FieldWorks\7.0\").GetValue(@"FwExeDir");
-					_fixitPathname = Path.Combine(codeDir, fixerName);
-				}
-				if (!File.Exists(_fixitPathname))
-				{
-					// Most developers have it here...
-					codeDir = @"C:\fwrepo\fw\Output\Debug";
-					_fixitPathname = Path.Combine(codeDir, fixerName);
-				}
-				if (!File.Exists(_fixitPathname))
-				{
-					// Or maybe a release build?
-					codeDir = @"C:\fwrepo\fw\Output\Release";
-					_fixitPathname = Path.Combine(codeDir, fixerName);
-				}
-			}
-			if (!File.Exists(_fixitPathname))
-				return false; // give up.
-
 			var process = new Process();
 			var startInfo = process.StartInfo;
-			startInfo.FileName = _fixitPathname;
-			startInfo.Arguments = "\"" + _fwdataPathname + "\"";
+			startInfo.FileName = "\"" + _fixitPathname.Replace("\"", null) + "\"";
+			startInfo.Arguments = "\"" + _fwdataPathname.Replace("\"", null) + "\"";
 			startInfo.CreateNoWindow = true; // don't need to bother the user with a dos prompt
 			startInfo.UseShellExecute = false;
 			startInfo.WorkingDirectory = Path.GetDirectoryName(_fixitPathname) ?? string.Empty;