+++ src/FLExBridge/Program.cs
@@ -1,13 +1,13 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel.Composition.Hosting;
 using System.IO;
 using System.Windows.Forms;
 using Chorus.VcsDrivers.Mercurial;
-using FLEx_ChorusPlugin.Properties;
 using Palaso.Reporting;
 using Palaso.UI.WindowsForms.HotSpot;
 using TriboroughBridge_ChorusPlugin;
+using TriboroughBridge_ChorusPlugin.Infrastructure;
+using TriboroughBridge_ChorusPlugin.Infrastructure.ActionHandlers;
 using TriboroughBridge_ChorusPlugin.Properties;
 
 namespace FLExBridge
@@ -20,6 +20,8 @@ namespace FLExBridge
 		[STAThread]
 		static void Main(string[] args)
 		{
+			//MessageBox.Show("Get ready to debug FB exe.");
+
 			// args are:
 			// -u username
 			// -p entire pathname to fwdata file including extension.
@@ -31,46 +33,74 @@ namespace FLExBridge
 			// since we don't actually use it directly, it does not show up when calling GetReferencedAssemblies on this assembly.
 			// But we need it to show up in that list so that ExceptionHandler.Init can install the intended PalasoUIWindowsForms
 			// exception handler.
-			new HotSpotProvider();
+			var hotspot = new HotSpotProvider();
+			hotspot.Dispose();
+
+			if (Settings.Default.CallUpgrade)
+			{
+				Settings.Default.Upgrade();
+				Settings.Default.CallUpgrade = false;
+			}
 
 			SetUpErrorHandling();
 			Application.EnableVisualStyles();
 			Application.SetCompatibleTextRenderingDefault(false);
 
-			// Is mercurial set up?
-			var readinessMessage = HgRepository.GetEnvironmentReadinessMessage("en");
-			if (!string.IsNullOrEmpty(readinessMessage))
-			{
-				MessageBox.Show(readinessMessage, CommonResources.kFLExBridge, MessageBoxButtons.OK, MessageBoxIcon.Asterisk);
-				return;
-			}
+			var options = CommandLineProcessor.ParseCommandLineArgs(args);
 
 			// An aggregate catalog that combines multiple catalogs
 			using (var catalog = new AggregateCatalog())
 			{
 				catalog.Catalogs.Add(new DirectoryCatalog(
-					Path.GetDirectoryName(Utilities.StripFilePrefix(typeof(BridgeTrafficCop).Assembly.CodeBase)),
+					Path.GetDirectoryName(Utilities.StripFilePrefix(typeof(ActionTypeHandlerRepository).Assembly.CodeBase)),
 					"*-ChorusPlugin.dll"));
 
 				// Create the CompositionContainer with the parts in the catalog
 				using (var container = new CompositionContainer(catalog))
 				{
-					var wantEndCall = false;
-					var options = ParseCommandLineArgs(args);
-					var bridgeTrafficCop = container.GetExportedValue<BridgeTrafficCop>();
+					var connHelper = container.GetExportedValue<FLExConnectionHelper>();
+					if (!connHelper.Init(options))
+						return;
+
+					// Is mercurial set up?
+					var readinessMessage = HgRepository.GetEnvironmentReadinessMessage("en");
+					if (!string.IsNullOrEmpty(readinessMessage))
+					{
+						MessageBox.Show(readinessMessage, CommonResources.kFLExBridge, MessageBoxButtons.OK, MessageBoxIcon.Asterisk);
+						return;
+					}
+
+					var l10Managers = Utilities.SetupLocalization(options);
+
 					try
 					{
-						bool showWindow;
-						wantEndCall = bridgeTrafficCop.StartWorking(options, out showWindow);
-						if (showWindow)
-							Application.Run(bridgeTrafficCop.MainForm);
+						var handlerRepository = container.GetExportedValue<ActionTypeHandlerRepository>();
+						var currentHandler = handlerRepository.GetHandler(options);
+						currentHandler.StartWorking(options);
+						var bridgeActionTypeHandlerShowWindow = currentHandler as IBridgeActionTypeHandlerShowWindow;
+						if (bridgeActionTypeHandlerShowWindow != null)
+						{
+							Application.Run(bridgeActionTypeHandlerShowWindow.MainForm);
+						}
+						var bridgeActionTypeHandlerCallEndWork = currentHandler as IBridgeActionTypeHandlerCallEndWork;
+						if (bridgeActionTypeHandlerCallEndWork != null)
+						{
+							bridgeActionTypeHandlerCallEndWork.EndWork();
+						}
+					}
+					catch
+					{
+						connHelper.SignalBridgeWorkComplete(false);
+						throw; // Re-throw the original exception, so the crash dlg has something to display.
 					}
 					finally
 					{
-						if (wantEndCall)
-							bridgeTrafficCop.EndWork(options);
-					}
+						foreach (var manager in l10Managers.Values)
+						{
+							manager.Dispose();
+						}
 
+					}
 				}
 			}
 			Settings.Default.Save();
@@ -78,43 +108,9 @@ namespace FLExBridge
 
 		private static void SetUpErrorHandling()
 		{
-			ErrorReport.EmailAddress = "fieldworksbridge@gmail.com";
+			ErrorReport.EmailAddress = Utilities.FlexBridgeEmailAddress;
 			ErrorReport.AddStandardProperties();
 			ExceptionHandler.Init();
-		}
-
-		static Dictionary<string, string> ParseCommandLineArgs(ICollection<string> args)
-		{
-			var options = new Dictionary<string, string>();
-			if (args != null && args.Count > 0)
-			{
-				string currentKey = null;
-				foreach (var arg in args)
-				{
-					//not all options are followed by input, so just add them as a key
-					if(arg.StartsWith("-") || arg.StartsWith("/"))
-					{
-						currentKey = arg.Trim();
-						options[currentKey] = null;
-					}
-					else //this is input which apparently follows an option, added it as the value in the dictionary
-					{
-						if (currentKey != null && options[currentKey] == null)
-						{
-							//this option goes with the flag that came before it
-							options[currentKey] = arg.Trim();
-						}
-						else //there was no flag before this option.
-						{
-							//This is an unparsable command line
-							Console.WriteLine(Resources.kInvalidCommandLineOptions);
-							//Signal FLEx or other apps
-							throw new ApplicationException("Invalid command line options.");
-						}
-					}
-				}
-			}
-			return options;
 		}
 	}
 }