+++ src/TheTurtleTests/Mocks/MockedProjectView.cs
@@ -1,0 +1,18 @@
+﻿using TheTurtle.View;
+
+namespace TheTurtleTests.Mocks
+{
+	internal class MockedProjectView : IProjectView
+	{
+		private readonly IExistingSystemView _existingSystemView = new MockedExistingSystemView();
+
+		#region Implementation of IProjectView
+
+		public IExistingSystemView ExistingSystemView
+		{
+			get { return _existingSystemView; }
+		}
+
+		#endregion
+	}
+}