+++ src/TriboroughBridge-ChorusPlugin/Controller/IMoveOldLiftRepositorController.cs
@@ -1,7 +1,0 @@
-﻿namespace TriboroughBridge_ChorusPlugin.Controller
-{
-	public interface IMoveOldLiftRepositorController : IObtainNewProjectController
-	{
-		void MoveRepoIfPresent();
-	}
-}