+++ src/TriboroughBridge-ChorusPlugin/Controller/JumpEventHandler.cs
@@ -1,4 +1,0 @@
-﻿namespace TriboroughBridge_ChorusPlugin.Controller
-{
-	public delegate void JumpEventHandler(object sender, JumpEventArgs e);
-}